Lecture03-04/notebooks/pytorch/5_scripting/pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "cifar10-classifier"
version         = "0.1.0"
description     = "Two-stage ResNet-18 transfer learning on CIFAR-10"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.1",
    "torchvision>=0.16",
    "torchmetrics>=1.2",
    "pyyaml>=6.0",
    "scikit-learn>=1.3",
    "numpy>=1.24",
    # serving
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",   # required for FastAPI file uploads
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
prepare-data = "scripts.prepare_data:main"
train        = "scripts.train:main"
evaluate     = "scripts.evaluate:main"
serve        = "scripts.serve:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version         = "3.10"
ignore_missing_imports = true
